requirements.txt
# YOLOv5 и PyTorch зависимости
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.23.5
pillow>=10.3.0
opencv-python-headless>=4.1.1
scipy>=1.4.1
PyYAML>=5.3.1
requests>=2.32.2
tqdm>=4.66.3
matplotlib>=3.3
pandas>=1.1.4
seaborn>=0.11.0
thop>=0.1.1
ultralytics>=8.2.34
setuptools>=70.0.0

# FastAPI и веб-сервер
fastapi
uvicorn[standard]

# Telegram бот
python-telegram-bot>=20.0
httpx

# Дополнительные утилиты
python-dotenv
psutil
gitpython>=3.1.30 